+++ ent/api/where.go
@@ -10,47 +10,47 @@ import (
 )
 
 // ID filters vertices based on their ID field.
-func ID(id int64) predicate.Api {
+func ID(id int32) predicate.Api {
 	return predicate.Api(sql.FieldEQ(FieldID, id))
 }
 
 // IDEQ applies the EQ predicate on the ID field.
-func IDEQ(id int64) predicate.Api {
+func IDEQ(id int32) predicate.Api {
 	return predicate.Api(sql.FieldEQ(FieldID, id))
 }
 
 // IDNEQ applies the NEQ predicate on the ID field.
-func IDNEQ(id int64) predicate.Api {
+func IDNEQ(id int32) predicate.Api {
 	return predicate.Api(sql.FieldNEQ(FieldID, id))
 }
 
 // IDIn applies the In predicate on the ID field.
-func IDIn(ids ...int64) predicate.Api {
+func IDIn(ids ...int32) predicate.Api {
 	return predicate.Api(sql.FieldIn(FieldID, ids...))
 }
 
 // IDNotIn applies the NotIn predicate on the ID field.
-func IDNotIn(ids ...int64) predicate.Api {
+func IDNotIn(ids ...int32) predicate.Api {
 	return predicate.Api(sql.FieldNotIn(FieldID, ids...))
 }
 
 // IDGT applies the GT predicate on the ID field.
-func IDGT(id int64) predicate.Api {
+func IDGT(id int32) predicate.Api {
 	return predicate.Api(sql.FieldGT(FieldID, id))
 }
 
 // IDGTE applies the GTE predicate on the ID field.
-func IDGTE(id int64) predicate.Api {
+func IDGTE(id int32) predicate.Api {
 	return predicate.Api(sql.FieldGTE(FieldID, id))
 }
 
 // IDLT applies the LT predicate on the ID field.
-func IDLT(id int64) predicate.Api {
+func IDLT(id int32) predicate.Api {
 	return predicate.Api(sql.FieldLT(FieldID, id))
 }
 
 // IDLTE applies the LTE predicate on the ID field.
-func IDLTE(id int64) predicate.Api {
+func IDLTE(id int32) predicate.Api {
 	return predicate.Api(sql.FieldLTE(FieldID, id))
 }
