+++ ent/groupmember.go
@@ -15,11 +15,11 @@ import (
 type GroupMember struct {
 	config `json:"-"`
 	// ID of the ent.
-	ID int `json:"id,omitempty"`
+	ID int32 `json:"id,omitempty"`
 	// GroupID holds the value of the "group_id" field.
-	GroupID int64 `json:"group_id,omitempty"`
+	GroupID int32 `json:"group_id,omitempty"`
 	// UserID holds the value of the "user_id" field.
-	UserID int64 `json:"user_id,omitempty"`
+	UserID int32 `json:"user_id,omitempty"`
 	// Role holds the value of the "role" field.
 	Role uint8 `json:"role,omitempty"`
 	// CreatedAt holds the value of the "created_at" field.
@@ -61,18 +61,18 @@ func (gm *GroupMember) assignValues(columns []string, values []any) error {
 			if !ok {
 				return fmt.Errorf("unexpected type %T for field id", value)
 			}
-			gm.ID = int(value.Int64)
+			gm.ID = int32(value.Int64)
 		case groupmember.FieldGroupID:
 			if value, ok := values[i].(*sql.NullInt64); !ok {
 				return fmt.Errorf("unexpected type %T for field group_id", values[i])
 			} else if value.Valid {
-				gm.GroupID = value.Int64
+				gm.GroupID = int32(value.Int64)
 			}
 		case groupmember.FieldUserID:
 			if value, ok := values[i].(*sql.NullInt64); !ok {
 				return fmt.Errorf("unexpected type %T for field user_id", values[i])
 			} else if value.Valid {
-				gm.UserID = value.Int64
+				gm.UserID = int32(value.Int64)
 			}
 		case groupmember.FieldRole:
 			if value, ok := values[i].(*sql.NullInt64); !ok {
